feat: Onion Archcitecture for Search Ingredients
- Implement architecture for searchIngredients
